Addition is a total function

+++ src/main/scala/com/exercise/BasicAlgebraicOperations.scala
@@ -2,7 +2,7 @@ package com.exercise
 
 object BasicAlgebraicOperations {
 
-  def add(a: Int, b: Int): Int = {
+  def add(a: BigInt, b: BigInt): BigInt = {
     a + b
   }
 }
